change profile page to server component
