+++ src/components/profile/profile.tsx
@@ -1,19 +1,30 @@
-/**
- * This code was generated by v0 by Vercel.
- * @see https://v0.dev/t/MVzmgZIaiJu
- */
+"use client";
+
 import ReactCardFlip from "react-card-flip";
 import ProfileFront from "./ProfileFront";
 import ProfileBack from "./ProfileBack";
+import { usePathname, useRouter, useSearchParams } from "next/navigation";
 
 type Props = {
   flip: string;
-  handleFront: (id: number) => void;
-  handleBack: () => void;
 };
 
 export function Profile(props: Props) {
-  const { flip, handleFront, handleBack } = props;
+  const { flip } = props;
+  const searchParams = useSearchParams();
+  const params = new URLSearchParams(searchParams);
+  const pathname = usePathname();
+  const { replace } = useRouter();
+
+  const handleFront = (flipId: number) => {
+    params.set("flip", flipId.toString());
+    replace(`${pathname}?${params.toString()}`, { scroll: false });
+  };
+
+  const handleBack = () => {
+    params.delete("flip");
+    replace(`${pathname}?${params.toString()}`, { scroll: false });
+  };
   return (
     <div className="grid sm:grid-cols-3 grid-cols-2 px-1 gap-4 mt-8">
       {[0, 1, 2, 3, 4, 5].map((index) => (
